+++ nxc/src/raw_memory_file.h
@@ -19,8 +19,8 @@ public:
 
 protected:
     // virtual bool _eof() const override;
-    virtual Result<void> _seek(SeekPos relative, size_t offset) override;
-    virtual Result<size_t> _tell() const override;
+    virtual Result<void> _seek(SeekPos relative, long offset) override;
+    virtual Result<long> _tell() const override;
 
     virtual Result<size_t> _read(void* buf, size_t n) override;
     virtual Result<size_t> _write(const void* buf, size_t n) override;
@@ -34,7 +34,7 @@ protected:
 private:
     RawMemoryPtr memory_;
     ConstRawMemoryPtr const_memory_;
-    size_t pos_;
+    long pos_;
     OpenMode mode_;
 };
